+++ R/plumber.R
@@ -143,7 +143,7 @@ Plumber <- R6Class(
     #' Mac OS X, port numbers smaller than 1025 require root privileges.
     #'
     #' This value does not need to be explicitly assigned. To explicitly set it, see [options_plumber()].
-    #' @param debug If `TRUE`, it will provide more insight into your API errors. Using this value will only last for the duration of the run. If a `$setDebug()` has not been called, `debug` will default to `interactive()` at `$run()` time. See `$setDebug()` for more details.
+    #' @param debug If `TRUE`, it will provide more insight into your API errors. Using this value will only last for the duration of the run. If a `$setDebug()` has not been called, `debug` will default to `FALSE` at `$run()` time. See `$setDebug()` for more details.
     #' @param swagger Deprecated. Please use `docs` instead. See `$setDocs(docs)` or `$setApiSpec()` for more customization.
     #' @param swaggerCallback An optional single-argument function that is
     #'   called back with the URL to an OpenAPI user interface when one becomes
@@ -159,7 +159,7 @@ Plumber <- R6Class(
     #' @importFrom rlang missing_arg
     run = function(
       host = '127.0.0.1',
-      port = getOption('plumber.port', NULL),
+      port = get_option_or_env('plumber.port', NULL),
       swagger = deprecated(),
       debug = missing_arg(),
       swaggerCallback = missing_arg(),
@@ -173,7 +173,7 @@ Plumber <- R6Class(
         stop("Plumber router `$run()` method should not be called while `plumb()`ing a file")
       }
 
-      ellipsis::check_dots_empty()
+      rlang::check_dots_empty()
 
       # Legacy support for RStudio pro products.
       # Checks must be kept for >= 2 yrs after plumber v1.0.0 release date
@@ -211,7 +211,7 @@ Plumber <- R6Class(
       port <- findPort(port)
 
       # Delay setting max size option. It could be set in `plumber.R`, which is after initialization
-      private$maxSize <- getOption('plumber.maxRequestSize', 0) #0  Unlimited
+      private$maxSize <- get_option_or_env('plumber.maxRequestSize', 0) #0  Unlimited
 
       # Delay the setting of swaggerCallback as long as possible.
       # An option could be set in `plumber.R`, which is after initialization
@@ -219,21 +219,20 @@ Plumber <- R6Class(
       swaggerCallback <-
         rlang::maybe_missing(swaggerCallback,
           rlang::maybe_missing(private$docs_callback,
-            getOption('plumber.docs.callback', getOption('plumber.swagger.url', NULL))
+            get_option_or_env('plumber.docs.callback', get_option_or_env('plumber.swagger.url', NULL))
           )
         )
 
-      # Delay the setting of debug as long as possible.
-      # The router could be made in an interactive setting and used in background process.
-      # Do not determine if interactive until run time
+      # Temporarily set the debug value
       prev_debug <- private$debug
       on.exit({
         private$debug <- prev_debug
       }, add = TRUE)
-      # Fix the debug value while running.
+      # Determine if the user should be informed about default behavior
+      inform_debug <- is.null(prev_debug) && rlang::is_missing(debug)
+
+      # Set debug value, defaulting to already set value (which returns FALSE if not set)
       self$setDebug(
-        # Order: Run method param, internally set value, is interactive()
-        # `$getDebug()` is dynamic given `setDebug()` has never been called.
         rlang::maybe_missing(debug, self$getDebug())
       )
 
@@ -265,6 +264,14 @@ Plumber <- R6Class(
           quiet = quiet
         )
         on.exit(unmount_docs(self, docs_info), add = TRUE)
+      }
+
+      if (!isTRUE(quiet) && inform_debug && rlang::is_interactive()) {
+        rlang::inform(
+          "Error reporting has been turned off by default. See `pr_set_debug()` for more details.\nTo disable this message, set a debug value.",
+          .frequency="once",
+          .frequency_id="pr_set_debug_message"
+        )
       }
 
       on.exit(private$runHooks("exit"), add = TRUE)
@@ -783,7 +790,7 @@ Plumber <- R6Class(
       # No endpoint could handle this request. 404
       notFoundStep <- function(...) {
 
-        if (isTRUE(getOption("plumber.trailingSlash", FALSE))) {
+        if (isTRUE(get_option_or_env("plumber.trailingSlash", FALSE))) {
           # Redirect to the slash route, if it exists
           path <- req$PATH_INFO
           # If the path does not end in a slash,
@@ -812,7 +819,7 @@ Plumber <- R6Class(
         # No trailing-slash route exists...
         # Try allowed verbs
 
-        if (isTRUE(getOption("plumber.methodNotAllowed", TRUE))) {
+        if (isTRUE(get_option_or_env("plumber.methodNotAllowed", TRUE))) {
           # Notify about allowed verbs
           if (is_405(req$pr, req$PATH_INFO, req$REQUEST_METHOD)) {
             res$status <- 405L
@@ -932,7 +939,7 @@ Plumber <- R6Class(
     #'  If using [options_plumber()], the value must be set before initializing your Plumber router.
     #' @param ... Arguments for the visual documentation. See each visual documentation package for further details.
     setDocs = function(
-      docs = getOption("plumber.docs", TRUE),
+      docs = get_option_or_env("plumber.docs", TRUE),
       ...
     ) {
       private$docs_info <- upgrade_docs_parameter(docs, ...)
@@ -947,7 +954,7 @@ Plumber <- R6Class(
     #' See also: [pr_set_docs_callback()]
     #' @param callback a callback function for taking action on the docs url. (Also accepts `NULL` values to disable the `callback`.)
     setDocsCallback = function(
-      callback = getOption('plumber.docs.callback', NULL)
+      callback = get_option_or_env('plumber.docs.callback', NULL)
     ) {
       # Use callback when defined
       if (!length(callback) || !is.function(callback)) {
@@ -962,15 +969,15 @@ Plumber <- R6Class(
     #'
     #' See also: `$getDebug()` and [pr_set_debug()]
     #' @param debug `TRUE` provides more insight into your API errors.
-    setDebug = function(debug = interactive()) {
+    setDebug = function(debug = FALSE) {
       stopifnot(length(debug) == 1)
       private$debug <- isTRUE(debug)
     },
-    #' @description Retrieve the `debug` value. If it has never been set, the result of `interactive()` will be used.
+    #' @description Retrieve the `debug` value. If it has never been set, it will return `FALSE`.
     #'
     #' See also: `$getDebug()` and [pr_set_debug()]
     getDebug = function() {
-      private$debug %||% default_debug()
+      private$debug %||% FALSE
     },
     #' @description Add a filter to plumber router
     #'
@@ -1331,13 +1338,6 @@ upgrade_docs_parameter <- function(docs, ...) {
     has_not_been_set = FALSE
   )
 }
-
-
-
-default_debug <- function() {
-  interactive()
-}
-
 
 urlHost <- function(scheme = "http", host, port, path = "", changeHostLocation = FALSE) {
   if (isTRUE(changeHostLocation)) {