+++ tests/testthat/test-plumber-run.R
@@ -77,8 +77,8 @@ test_that("`swaggerCallback` can be set by option after the pr is created", {
 
   with_options(
     list(
-      plumber.swagger.url = getOption("plumber.swagger.url"),
-      plumber.docs.callback = getOption("plumber.docs.callback")
+      plumber.swagger.url = get_option_or_env("plumber.swagger.url"),
+      plumber.docs.callback = get_option_or_env("plumber.docs.callback")
     ),
     {
       # set option after init
@@ -91,46 +91,5 @@ test_that("`swaggerCallback` can be set by option after the pr is created", {
   )
   # m is used
   mockery::expect_called(m, 1)
-
-})
-
-
-### Test does not work as expected with R6 objects.
-test_that("`debug` is not set until runtime", {
-  skip_if_not_installed("mockery", "0.4.2")
-
-  m <- mockery::mock(TRUE, cycle = TRUE)
-  # https://github.com/r-lib/testthat/issues/734#issuecomment-377367516
-  # > It should work if you fully qualify the function name (include the pkgname)
-  with_mock("plumber:::default_debug" = m, {
-    root <- pr()
-    root$getDebug()
-    mockery::expect_called(m, 1)
-
-    with_interrupt({
-      root %>% pr_run(quiet = TRUE)
-    })
-    # increase by 1
-    mockery::expect_called(m, 2)
-
-    # listen to set value
-    with_interrupt({
-      root %>%
-        pr_set_debug(TRUE) %>%
-        pr_run(quiet = TRUE)
-    })
-    # not updated. stay at 2
-    mockery::expect_called(m, 2)
-
-    # listen to run value
-    with_interrupt({
-      root %>%
-        pr_run(debug = FALSE, quiet = TRUE)
-    })
-    # not updated. stay at 2
-    mockery::expect_called(m, 2)
-
-    # TODO test that run(debug=) has preference over pr_set_debug()
-  })
 
 })